Minor updates and tests

+++ Linear/Main.cpp
@@ -10,8 +10,8 @@ int main()
 {
 	Stopwatch<std::chrono::milliseconds> s;
 
-	//*
-	Eigen::Matrix<float, -1, -1, Eigen::RowMajor> mat1;
+	/*
+	Eigen::Matrix<double, -1, -1, Eigen::RowMajor> mat1;
 	Eigen::Matrix<int, -1, -1, Eigen::RowMajor> mat2;
 	mat1.resize(2500, 2500);
 	mat2.resize(25, 25);
@@ -34,7 +34,7 @@ int main()
 
 	//Eigen::Matrix<int, -1, -1, Eigen::RowMajor>mm = mat1 * mat2;
 	static constexpr int numTests = 20;
-	Matrix<float> tt(2500, 2500);
+	Matrix<double> tt(2500, 2500);
 	s.start();
 	for (int i = 0; i < numTests; ++i)
 		tt = tt * tt;
